#12-modal-refactor
Refactored modal to global model renderer controlled by one store.Added confirm dialogs.

+++ frontend/src/components/TaskDocumentsList/TaskDocumentsList.tsx
@@ -58,9 +58,11 @@ const TaskDocumentsListItem = observer(
           <Tooltip text="Delete document">
             <button
               onClick={() =>
-                rootStore.taskStore.deleteTaskFilesById(
-                  String(file.id),
-                  file.taskId
+                rootStore.modalStore.showConfirm("Are you sure?", () =>
+                  rootStore.taskStore.deleteTaskFilesById(
+                    String(file.id),
+                    file.taskId
+                  )
                 )
               }
               className={styles.deleteButton}
